requirements.txt
attrs==21.4.0
cffi==1.15.0
colorama==0.4.4
execnet==1.9.0
future-fstrings==1.2.0
iniconfig==1.1.1
msgpack==1.0.3
networkx==2.6.3
packaging==21.3
pluggy==1.0.0
py==1.11.0
py-algorand-sdk==1.15.0
pycparser==2.21
pycryptodomex==3.12.0
pydantic==1.9.0
PyNaCl==1.5.0
pyparsing==3.0.6
pyteal==0.13.0
-e git+https://github.com/algorand/pyteal-utils.git@main#egg=pyteal_utils
pytest==6.2.5
pytest-depends==1.0.1
pytest-forked==1.4.0
pytest-xdist==2.5.0
PyYAML==6.0
toml==0.10.2
typing-extensions==4.0.1
